[~] - 'io.qase.configuration' was renamed to 'io.qase.guice.module'
