+++ qase-cucumber3-jvm/src/main/java/io/qase/cucumber3/QaseEventListener.java
@@ -23,7 +23,7 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static io.qase.api.utils.IntegrationUtils.getStacktrace;
-import static io.qase.configuration.QaseModule.INJECTOR;
+import static io.qase.guice.module.QaseModule.INJECTOR;
 
 public class QaseEventListener implements Formatter {
